move evaluator, score_evaluator, test_metric back to the backend
