+++ sdk/src/rhesis/sdk/metrics/__init__.py
@@ -3,7 +3,6 @@
 from .base import BaseMetric, MetricConfig, MetricResult
 from .config.loader import MetricConfigLoader
 from .constants import OPERATOR_MAP, VALID_OPERATORS_BY_SCORE_TYPE, ScoreType, ThresholdOperator
-from .evaluator import MetricEvaluator as Evaluator
 from .factory import MetricFactory
 
 # Lazy import to avoid circular dependencies
@@ -27,7 +26,6 @@ from .rhesis import (  # Re-export Rhesis metrics
     RhesisMetricFactory,
     RhesisPromptMetric,
 )
-from .score_evaluator import ScoreEvaluator
 from .utils import diagnose_invalid_metric, run_evaluation
 
 __all__ = [
@@ -37,10 +35,6 @@ __all__ = [
     "MetricResult",
     "MetricConfigLoader",
     "MetricFactory",
-    # Evaluation
-    "Evaluator",
-    "ScoreEvaluator",
-    "run_evaluation",
     # Types and utilities
     "ScoreType",
     "ThresholdOperator",
